requirements.txt
langchain==0.0.125
streamlit==1.20.0
streamlit-chat==0.0.2.2
openai==0.27.4
beautifulsoup4==4.12.2
googlesearch-python==1.2.3
google-crc32c==1.5.0
requests-oauthlib==1.3.0
jsonschema==4.17.3